BiletDB
This database application offers a platform that enables users to purchase and book tickets for various events such as events, concerts, theatre plays, cinema sessions. The application works integrated with SQL database. The C# application stores user credentials, contact information and account details of users registered in the system in the SQL database.